requirements.txt
ultralytics
opencv-python
easyocr
pyyaml
numpy
torch
torchvision
